Project Download change

+++ ai_workspace/api_views.py
@@ -1729,13 +1729,13 @@ def previously_created_steps(request):
 @api_view(["GET"])
 def project_download(request,project_id):
     pr = Project.objects.get(id=project_id)
-    #if os.path.exists(os.path.join(pr.project_dir_path,'source')):
-    shutil.make_archive(pr.project_name, 'zip', pr.project_dir_path + '/source')
-    res = download_file(pr.project_name+'.zip')
-    os.remove(pr.project_name+'.zip')
-    return res
-    # else:
-    #     return Response({'msg':'something went wrong'})
+    if os.path.exists(os.path.join(pr.project_dir_path,'source')):
+        shutil.make_archive(pr.project_name, 'zip', pr.project_dir_path + '/source')
+        res = download_file(pr.project_name+'.zip')
+        os.remove(pr.project_name+'.zip')
+        return res
+    else:
+        return Response({'msg':'something went wrong'})
 
 class ShowMTChoices(APIView):
     # permission_classes = [IsAuthenticated]
